adding distributed reinforcement learning based on the ray framework
