+++ benchmarks/Test/Test_CartPole-v1_DQN_ray_20230330-215646/logs/log.txt
@@ -1,0 +1,4 @@
+2023-03-30 21:56:48 - r - INFO: - n_states: 4, n_actions: 2
+2023-03-30 21:56:50 - r - INFO: - Start testing!
+2023-03-30 21:56:50 - r - INFO: - Env: CartPole-v1, Algorithm: DQN, Device: cpu
+2023-03-30 21:56:55 - r - INFO: - Finish testing!
