+++ benchmarks/Train/Train_CartPole-v1_DQN_ray_20230330-212958/logs/log.txt
@@ -1,0 +1,4 @@
+2023-03-30 21:30:00 - r - INFO: - n_states: 4, n_actions: 2
+2023-03-30 21:30:01 - r - INFO: - Start training!
+2023-03-30 21:30:01 - r - INFO: - Env: CartPole-v1, Algorithm: DQN, Device: cpu
+2023-03-30 21:38:45 - r - INFO: - Finish training!
