+++ benchmarks/Train/Train_CartPole-v1_DQN_ray_20230330-212958/results/res.csv
@@ -1,0 +1,401 @@
+episodes,rewards
+0,22.0
+1,34.0
+2,22.0
+3,34.0
+4,27.0
+5,13.0
+6,17.0
+7,18.0
+8,30.0
+9,16.0
+10,35.0
+11,12.0
+12,15.0
+13,10.0
+14,18.0
+15,10.0
+16,16.0
+17,14.0
+18,23.0
+19,23.0
+20,18.0
+21,13.0
+22,31.0
+23,10.0
+24,15.0
+25,9.0
+26,18.0
+27,17.0
+28,36.0
+29,30.0
+30,21.0
+31,16.0
+32,21.0
+33,28.0
+34,23.0
+35,26.0
+36,11.0
+37,11.0
+38,16.0
+39,14.0
+40,17.0
+41,15.0
+42,17.0
+43,10.0
+44,9.0
+45,13.0
+46,12.0
+47,14.0
+48,14.0
+49,14.0
+50,10.0
+51,9.0
+52,12.0
+53,10.0
+54,10.0
+55,9.0
+56,10.0
+57,9.0
+58,12.0
+59,9.0
+60,11.0
+61,13.0
+62,12.0
+63,17.0
+64,15.0
+65,11.0
+66,9.0
+67,12.0
+68,9.0
+69,14.0
+70,9.0
+71,11.0
+72,14.0
+73,12.0
+74,10.0
+75,9.0
+76,20.0
+77,15.0
+78,10.0
+79,17.0
+80,17.0
+81,12.0
+82,11.0
+83,17.0
+84,16.0
+85,14.0
+86,12.0
+87,18.0
+88,11.0
+89,13.0
+90,10.0
+91,10.0
+92,9.0
+93,9.0
+94,10.0
+95,9.0
+96,10.0
+97,9.0
+98,10.0
+99,9.0
+100,11.0
+101,9.0
+102,10.0
+103,9.0
+104,12.0
+105,11.0
+106,10.0
+107,9.0
+108,12.0
+109,9.0
+110,11.0
+111,11.0
+112,116.0
+113,99.0
+114,61.0
+115,42.0
+116,29.0
+117,31.0
+118,21.0
+119,28.0
+120,16.0
+121,19.0
+122,27.0
+123,20.0
+124,24.0
+125,19.0
+126,22.0
+127,16.0
+128,18.0
+129,23.0
+130,20.0
+131,23.0
+132,25.0
+133,26.0
+134,19.0
+135,32.0
+136,18.0
+137,20.0
+138,24.0
+139,20.0
+140,29.0
+141,25.0
+142,51.0
+143,30.0
+144,31.0
+145,60.0
+146,40.0
+147,71.0
+148,78.0
+149,46.0
+150,29.0
+151,29.0
+152,46.0
+153,29.0
+154,43.0
+155,27.0
+156,27.0
+157,48.0
+158,31.0
+159,58.0
+160,31.0
+161,37.0
+162,55.0
+163,35.0
+164,43.0
+165,61.0
+166,50.0
+167,66.0
+168,65.0
+169,66.0
+170,63.0
+171,99.0
+172,55.0
+173,50.0
+174,30.0
+175,53.0
+176,32.0
+177,34.0
+178,51.0
+179,31.0
+180,43.0
+181,32.0
+182,48.0
+183,35.0
+184,31.0
+185,29.0
+186,23.0
+187,34.0
+188,32.0
+189,42.0
+190,32.0
+191,37.0
+192,58.0
+193,32.0
+194,52.0
+195,31.0
+196,28.0
+197,39.0
+198,34.0
+199,58.0
+200,36.0
+201,39.0
+202,106.0
+203,92.0
+204,100.0
+205,80.0
+206,94.0
+207,125.0
+208,90.0
+209,180.0
+210,200.0
+211,148.0
+212,160.0
+213,200.0
+214,146.0
+215,81.0
+216,123.0
+217,200.0
+218,197.0
+219,156.0
+220,200.0
+221,200.0
+222,200.0
+223,200.0
+224,200.0
+225,200.0
+226,200.0
+227,200.0
+228,200.0
+229,200.0
+230,200.0
+231,200.0
+232,200.0
+233,182.0
+234,164.0
+235,147.0
+236,172.0
+237,166.0
+238,200.0
+239,200.0
+240,200.0
+241,200.0
+242,200.0
+243,191.0
+244,179.0
+245,182.0
+246,200.0
+247,200.0
+248,200.0
+249,200.0
+250,191.0
+251,178.0
+252,166.0
+253,170.0
+254,200.0
+255,200.0
+256,200.0
+257,200.0
+258,200.0
+259,200.0
+260,200.0
+261,200.0
+262,185.0
+263,185.0
+264,200.0
+265,200.0
+266,199.0
+267,200.0
+268,200.0
+269,200.0
+270,200.0
+271,192.0
+272,190.0
+273,189.0
+274,193.0
+275,200.0
+276,200.0
+277,200.0
+278,200.0
+279,190.0
+280,186.0
+281,200.0
+282,200.0
+283,191.0
+284,195.0
+285,200.0
+286,200.0
+287,200.0
+288,200.0
+289,161.0
+290,169.0
+291,200.0
+292,200.0
+293,200.0
+294,200.0
+295,128.0
+296,129.0
+297,133.0
+298,138.0
+299,175.0
+300,156.0
+301,125.0
+302,127.0
+303,169.0
+304,171.0
+305,136.0
+306,147.0
+307,140.0
+308,199.0
+309,197.0
+310,166.0
+311,200.0
+312,190.0
+313,198.0
+314,183.0
+315,192.0
+316,200.0
+317,200.0
+318,200.0
+319,198.0
+320,195.0
+321,192.0
+322,186.0
+323,200.0
+324,200.0
+325,200.0
+326,200.0
+327,200.0
+328,149.0
+329,160.0
+330,200.0
+331,140.0
+332,200.0
+333,200.0
+334,196.0
+335,185.0
+336,195.0
+337,200.0
+338,200.0
+339,121.0
+340,200.0
+341,198.0
+342,132.0
+343,196.0
+344,198.0
+345,200.0
+346,200.0
+347,200.0
+348,200.0
+349,200.0
+350,200.0
+351,200.0
+352,198.0
+353,200.0
+354,200.0
+355,200.0
+356,200.0
+357,200.0
+358,193.0
+359,160.0
+360,200.0
+361,192.0
+362,161.0
+363,153.0
+364,153.0
+365,152.0
+366,154.0
+367,159.0
+368,165.0
+369,199.0
+370,200.0
+371,200.0
+372,200.0
+373,200.0
+374,200.0
+375,200.0
+376,200.0
+377,200.0
+378,176.0
+379,176.0
+380,174.0
+381,200.0
+382,200.0
+383,200.0
+384,200.0
+385,165.0
+386,158.0
+387,169.0
+388,159.0
+389,194.0
+390,169.0
+391,140.0
+392,132.0
+393,124.0
+394,122.0
+395,200.0
+396,200.0
+397,163.0
+398,166.0
+399,158.0
