+++ main.py
@@ -11,6 +11,8 @@ import yaml
 from pathlib import Path
 import datetime
 import gym
+import ray
+from ray.util.queue import Queue
 import torch.multiprocessing as mp
 from torch.utils.tensorboard import SummaryWriter 
 from config.config import GeneralConfig
@@ -227,6 +229,48 @@ class Main(object):
         plot_rewards(rewards,
                      title=f"{cfg.mode.lower()}ing curve on {cfg.device} of {cfg.algo_name} for {cfg.env_name}",
                      fpath=cfg.res_dir)
+        
+    def ray_run(self,cfg):
+        ''' 使用Ray并行化强化学习算法
+        '''
+        ray.init()
+        envs = self.envs_config(cfg)  # configure environment
+        agent_mod = __import__(f"algos.{cfg.algo_name}.agent", fromlist=['Agent'])
+        agent_mod = __import__(f"algos.{cfg.algo_name}.agent", fromlist=['ShareAgent'])
+        share_agent = agent_mod.ShareAgent.remote(cfg)  # create agent
+        local_agents = [agent_mod.Agent(cfg) for _ in range(cfg.n_workers)]
+        worker_mod = __import__(f"algos.{cfg.algo_name}.trainer", fromlist=['WorkerRay'])
+        if cfg.load_checkpoint:
+            ray.get(share_agent.load_model.remote(f"tasks/{cfg.load_path}/models"))
+            for local_agent in local_agents:
+                local_agent.load_model(f"tasks/{cfg.load_path}/models")
+        self.logger.info(f"Start {cfg.mode}ing!")
+        self.logger.info(f"Env: {cfg.env_name}, Algorithm: {cfg.algo_name}, Device: {cfg.device}")
+        global_r_que = Queue()
+        # print(f'cfg.n_workers:{cfg.n_workers}')
+        workerrays = [worker_mod.WorkerRay.remote(cfg, i, share_agent, envs[i], local_agents[i], global_r_que) for i in range(cfg.n_workers)]
+        task_ids = [w.run.remote() for w in workerrays]
+        # 等待所有任务完成, 注意：ready_ids, task_ids变量不能随意改。
+        while len(task_ids) > 0:
+            ready_ids, task_ids = ray.wait(task_ids)
+        rewards = []
+        global_r_que_length = len(global_r_que)
+        for _ in range(global_r_que_length):
+            rewards.append(global_r_que.get())
+        # sorted_dict_list形如[{episode：reward}, {episode：reward} ...]。将{episode：reward}按照episode顺序排序
+        sorted_dict_list = sorted(rewards, key=lambda x: list(x.keys())[0])
+        # 取出value，形成数组
+        rewards = [list(d.values())[0] for d in sorted_dict_list]
+
+        ray.shutdown()
+        self.logger.info(f"Finish {cfg.mode}ing!")
+        res_dic = {'episodes': range(len(rewards)), 'rewards': rewards}
+        save_results(res_dic, cfg.res_dir)  # save results
+        save_cfgs(self.cfgs, cfg.task_dir)  # save config
+        plot_rewards(rewards,
+                     title=f"{cfg.mode.lower()}ing curve of {cfg.algo_name} for {cfg.env_name} with {cfg.n_workers} {cfg.device}",
+                     fpath=cfg.res_dir)
+
     def run(self) -> None:
         self.get_default_cfg()  # get default config
         self.process_yaml_cfg()  # process yaml config
